wip: update meeting screen
